Loader moved out of Button Sprucebot.components contains Loader and Typeography
