+++ lib/Sprucebot.js
@@ -6,6 +6,8 @@ Object.defineProperty(exports, '__esModule', {
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj['default'] = obj; return newObj; } }
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
 var _componentsAvatarAvatar = require('./components/Avatar/Avatar');
@@ -40,9 +42,13 @@ var _componentsSwitchSwitch = require('./components/Switch/Switch');
 
 var _componentsSwitchSwitch2 = _interopRequireDefault(_componentsSwitchSwitch);
 
+var _componentsLoaderLoader = require('./components/Loader/Loader');
+
+var _componentsLoaderLoader2 = _interopRequireDefault(_componentsLoaderLoader);
+
 var _componentsTypeographyTypeography = require('./components/Typeography/Typeography');
 
-var _componentsTypeographyTypeography2 = _interopRequireDefault(_componentsTypeographyTypeography);
+var Typeography = _interopRequireWildcard(_componentsTypeographyTypeography);
 
 var Sprucebot = {
     components: _extends({
@@ -53,8 +59,9 @@ var Sprucebot = {
         Container: _componentsContainerContainer2['default'],
         Input: _componentsInputInput2['default'],
         Pre: _componentsPrePre2['default'],
-        Switch: _componentsSwitchSwitch2['default']
-    }, _componentsTypeographyTypeography2['default'])
+        Switch: _componentsSwitchSwitch2['default'],
+        Loader: _componentsLoaderLoader2['default']
+    }, Typeography)
 };
 exports['default'] = Sprucebot;
 var components = Sprucebot.components;
