+++ lib/components/Button/Button.js
@@ -28,9 +28,9 @@ var _propTypes = require('prop-types');
 
 var _propTypes2 = _interopRequireDefault(_propTypes);
 
-var _Loader = require('./Loader');
+var _LoaderLoader = require('../Loader/Loader');
 
-var _Loader2 = _interopRequireDefault(_Loader);
+var _LoaderLoader2 = _interopRequireDefault(_LoaderLoader);
 
 // TODO refactor into styled component
 
@@ -94,7 +94,7 @@ var Button = (function (_Component) {
 			return _react2['default'].createElement(
 				Tag,
 				_extends({ className: 'btn__' + btnClass + ' ' + (className || '') }, props),
-				busy ? _react2['default'].createElement(_Loader2['default'], { blue: false, fullWidth: false }) : _react2['default'].createElement(
+				busy ? _react2['default'].createElement(_LoaderLoader2['default'], { blue: false, fullWidth: false }) : _react2['default'].createElement(
 					_reactEmojione2['default'],
 					null,
 					children
