+++ lib/components/Loader/Loader.js
@@ -35,10 +35,10 @@ var Loader = (function (_Component) {
 		key: 'render',
 		value: function render() {
 			var _props = this.props;
-			var blue = _props.blue;
+			var dark = _props.dark;
 			var fullWidth = _props.fullWidth;
 
-			var dotClassName = blue ? 'loader_dot_blue' : 'loader_dot';
+			var dotClassName = dark ? 'loader_dot_dark' : 'loader_dot';
 			var fullWidthStyle = fullWidth ? { display: 'block', margin: '20px', textAlign: 'center' } : {};
 
 			return _react2['default'].createElement(
@@ -57,12 +57,12 @@ var Loader = (function (_Component) {
 exports['default'] = Loader;
 
 Loader.propTypes = {
-	blue: _propTypes2['default'].bool,
+	dark: _propTypes2['default'].bool,
 	fullWidth: _propTypes2['default'].bool
 };
 
 Loader.defaultProps = {
-	blue: true,
+	dark: true,
 	fullWidth: true
 };
 module.exports = exports['default'];
